Backend2/infrastructure/lambda/requirements.txt
jsonpickle
aws_xray_sdk
boto3
pyjwt